Move the pulse generator out to a helper so it can be reused
